# replace encoding/json with goccy/go-json (#2421 expanded)Fast json