--- a/_examples/benchmarking/.gqlgen.yml
+++ b/_examples/benchmarking/.gqlgen.yml
@@ -1,0 +1,4 @@
+exec:
+  filename: generated/exec.go
+model:
+  filename: models/generated.go