--- a/_examples/benchmarking/readme.md
+++ b/_examples/benchmarking/readme.md
@@ -1,0 +1,14 @@
+### starwars example
+
+This server demonstrates a few advanced features of graphql:
+ - connections
+ - unions
+ - interfaces
+ - enums
+
+to run this server
+```bash
+go run ./server/server.go
+```
+
+and open http://localhost:8080 in your browser