--- a/graphql/any.go
+++ b/graphql/any.go
@@ -1,7 +1,7 @@
 package graphql
 
 import (
-	"encoding/json"
+	"github.com/goccy/go-json"
 	"io"
 )