--- a/graphql/coercion_test.go
+++ b/graphql/coercion_test.go
@@ -1,7 +1,7 @@
 package graphql
 
 import (
-	"encoding/json"
+	"github.com/goccy/go-json"
 	"testing"
 
 	"github.com/stretchr/testify/assert"