--- a/graphql/float.go
+++ b/graphql/float.go
@@ -2,9 +2,9 @@ package graphql
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/goccy/go-json"
 	"io"
 	"math"
 	"strconv"