--- a/graphql/handler/transport/error.go
+++ b/graphql/handler/transport/error.go
@@ -1,10 +1,10 @@
 package transport
 
 import (
-	"encoding/json"
 	"fmt"
 	"net/http"
 
+	"github.com/goccy/go-json"
 	"github.com/vektah/gqlparser/v2/gqlerror"
 
 	"github.com/99designs/gqlgen/graphql"