--- a/graphql/handler/transport/http_form_multipart.go
+++ b/graphql/handler/transport/http_form_multipart.go
@@ -1,7 +1,7 @@
 package transport
 
 import (
-	"encoding/json"
+	"github.com/goccy/go-json"
 	"io"
 	"mime"
 	"net/http"