--- a/graphql/handler/transport/util.go
+++ b/graphql/handler/transport/util.go
@@ -2,8 +2,9 @@ package transport
 
 import (
 	"fmt"
-	"github.com/goccy/go-json"
 	"io"
+
+	"github.com/goccy/go-json"
 
 	"github.com/vektah/gqlparser/v2/gqlerror"