--- a/graphql/handler/transport/websocket_graphql_transport_ws.go
+++ b/graphql/handler/transport/websocket_graphql_transport_ws.go
@@ -2,6 +2,7 @@ package transport
 
 import (
 	"fmt"
+
 	"github.com/goccy/go-json"
 
 	"github.com/gorilla/websocket"