--- a/graphql/handler/transport/websocket_graphqlws.go
+++ b/graphql/handler/transport/websocket_graphqlws.go
@@ -2,6 +2,7 @@ package transport
 
 import (
 	"fmt"
+
 	"github.com/goccy/go-json"
 
 	"github.com/gorilla/websocket"