--- a/graphql/handler/transport/websocket_subprotocol.go
+++ b/graphql/handler/transport/websocket_subprotocol.go
@@ -1,8 +1,8 @@
 package transport
 
 import (
-	"encoding/json"
 	"errors"
+	"github.com/goccy/go-json"
 
 	"github.com/gorilla/websocket"
 )