--- a/graphql/handler/transport/websocket_subprotocol.go
+++ b/graphql/handler/transport/websocket_subprotocol.go
@@ -2,6 +2,7 @@ package transport
 
 import (
 	"errors"
+
 	"github.com/goccy/go-json"
 
 	"github.com/gorilla/websocket"