--- a/graphql/id.go
+++ b/graphql/id.go
@@ -1,8 +1,8 @@
 package graphql
 
 import (
-	"encoding/json"
 	"fmt"
+	"github.com/goccy/go-json"
 	"io"
 	"strconv"
 )