--- a/graphql/int.go
+++ b/graphql/int.go
@@ -1,8 +1,8 @@
 package graphql
 
 import (
-	"encoding/json"
 	"fmt"
+	"github.com/goccy/go-json"
 	"io"
 	"strconv"
 )