--- a/graphql/int_test.go
+++ b/graphql/int_test.go
@@ -1,8 +1,9 @@
 package graphql
 
 import (
-	"github.com/goccy/go-json"
 	"testing"
+
+	"github.com/goccy/go-json"
 
 	"github.com/stretchr/testify/assert"
 )