--- a/graphql/omittable_test.go
+++ b/graphql/omittable_test.go
@@ -1,8 +1,9 @@
 package graphql
 
 import (
-	"github.com/goccy/go-json"
 	"testing"
+
+	"github.com/goccy/go-json"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"