--- a/graphql/response.go
+++ b/graphql/response.go
@@ -2,8 +2,8 @@ package graphql
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
+	"github.com/goccy/go-json"
 
 	"github.com/vektah/gqlparser/v2/ast"
 	"github.com/vektah/gqlparser/v2/gqlerror"