--- a/graphql/response.go
+++ b/graphql/response.go
@@ -3,6 +3,7 @@ package graphql
 import (
 	"context"
 	"fmt"
+
 	"github.com/goccy/go-json"
 
 	"github.com/vektah/gqlparser/v2/ast"