--- a/graphql/string_test.go
+++ b/graphql/string_test.go
@@ -1,8 +1,9 @@
 package graphql
 
 import (
-	"github.com/goccy/go-json"
 	"testing"
+
+	"github.com/goccy/go-json"
 
 	"github.com/stretchr/testify/assert"
 )