--- a/graphql/uint.go
+++ b/graphql/uint.go
@@ -3,9 +3,10 @@ package graphql
 import (
 	"errors"
 	"fmt"
-	"github.com/goccy/go-json"
 	"io"
 	"strconv"
+
+	"github.com/goccy/go-json"
 )
 
 func MarshalUint(i uint) Marshaler {