--- a/graphql/uint_test.go
+++ b/graphql/uint_test.go
@@ -1,7 +1,7 @@
 package graphql
 
 import (
-	"encoding/json"
+	"github.com/goccy/go-json"
 	"math"
 	"testing"