This is a revival of pr #2421.

I switched over to https://github.com/goccy/go-json instead of encoding/json. Previously it wasn't accepted because it only showed a ~5% improvement in performance in the starwars test suite. I added a test that handles larger json requests  and responses (_examples/benchmarking) and in those cases the improvement was better than 5%.
```goos: darwin
goarch: arm64
pkg: github.com/99designs/gqlgen/_examples/benchmarking
                                                                      │ libjson.txt  │            goccyjson.txt            │
                                                                      │    sec/op    │   sec/op     vs base                │
QueriesOfVariableSizes/input_size:_100000_output_size_100000-10          1.367m ± 0%   1.004m ± 1%  -26.60% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_1000000_output_size_1000000-10       13.176m ± 1%   8.746m ± 1%  -33.62% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_10000000_output_size_1000000-10       61.62m ± 1%   28.30m ± 1%  -54.07% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_10000000_output_size_10000000-10     127.81m ± 0%   83.67m ± 0%  -34.54% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_100000000_output_size_100000000-10   1293.4m ± 3%   822.3m ± 1%  -36.42% (p=0.000 n=10)
geomean                                                                  44.95m        27.96m       -37.79%

                                                                      │ libjson.txt  │            goccyjson.txt             │
                                                                      │     B/op     │     B/op      vs base                │
QueriesOfVariableSizes/input_size:_100000_output_size_100000-10         1.491Mi ± 0%   1.648Mi ± 1%  +10.50% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_1000000_output_size_1000000-10       14.94Mi ± 1%   17.06Mi ± 1%  +14.19% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_10000000_output_size_1000000-10      134.4Mi ± 0%   135.3Mi ± 0%   +0.72% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_10000000_output_size_10000000-10     160.6Mi ± 2%   186.8Mi ± 0%  +16.34% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_100000000_output_size_100000000-10   1.754Gi ± 0%   1.848Gi ± 0%   +5.31% (p=0.000 n=10)
geomean                                                                 61.27Mi        66.94Mi        +9.26%

                                                                      │ libjson.txt │           goccyjson.txt            │
                                                                      │  allocs/op  │ allocs/op   vs base                │
QueriesOfVariableSizes/input_size:_100000_output_size_100000-10          214.0 ± 0%   213.0 ± 0%   -0.47% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_1000000_output_size_1000000-10        239.5 ± 0%   235.0 ± 0%   -1.88% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_10000000_output_size_1000000-10       259.0 ± 0%   253.0 ± 0%   -2.32% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_10000000_output_size_10000000-10      269.0 ± 1%   262.0 ± 0%   -2.60% (p=0.000 n=10)
QueriesOfVariableSizes/input_size:_100000000_output_size_100000000-10    419.0 ± 0%   336.5 ± 0%  -19.69% (p=0.000 n=10)
geomean                                                                  272.3        256.8        -5.69%
```

As you can see here the runtime improvement is 37% on average and the memory allocations are less frequent averaging 5% less allocations and the memory usage increased by less than 10% on average.

If you prefer i can make this a configurable value, it'll be a bit tricky and i'll probably have to use some sort of global variable.

I have:
 - [ ] Added tests covering the bug / feature (see [testing](https://github.com/99designs/gqlgen/blob/master/TESTING.md))
 - [ ] Updated any relevant documentation (see [docs](https://github.com/99designs/gqlgen/tree/master/docs/content))