@adomaskizogian 2024-09-13
DO NOT MERGE THIS
I REPEAT
DO NOT EVER MERGE THIS

goccy/go-json IS NOT A DROP IN REPLACEMENT FOR STDLIB `encoding/json`.
It is extremely broken implementation and the bugs are not some edge cases but very easy to reproduce ones.

https://github.com/goccy/go-json/issues/510
https://github.com/goccy/go-json/issues/512
https://github.com/goccy/go-json/issues/519