@giautm 2024-09-16
It should be configurable, to generate code depends on the end-user choose, but I prefer standard library for all of cases.